Refactor/turn use case

+++ game-api/src/main/java/ca/ulaval/glo4002/game/application/turn/TurnUseCase.java
@@ -1,47 +1,44 @@
 package ca.ulaval.glo4002.game.application.turn;
 
-import ca.ulaval.glo4002.game.controllers.turn.dtos.TurnAssembler;
+import ca.ulaval.glo4002.game.application.manager.ZooManager;
 import ca.ulaval.glo4002.game.domain.actions.Action;
 import ca.ulaval.glo4002.game.domain.actions.ActionRepository;
 import ca.ulaval.glo4002.game.domain.dinosaur.Dinosaur;
 import ca.ulaval.glo4002.game.domain.dinosaur.DinosaurBaby;
 import ca.ulaval.glo4002.game.domain.dinosaur.DinosaurRepository;
-import ca.ulaval.glo4002.game.domain.dinosaur.enums.DietType;
 import ca.ulaval.glo4002.game.domain.resources.Burger;
 import ca.ulaval.glo4002.game.domain.resources.ResourceRepository;
 import ca.ulaval.glo4002.game.domain.resources.Salad;
 import ca.ulaval.glo4002.game.domain.resources.Water;
-import ca.ulaval.glo4002.game.controllers.turn.dtos.TurnDto;
 import ca.ulaval.glo4002.game.domain.turn.Turn;
 import ca.ulaval.glo4002.game.domain.turn.TurnFactory;
 import ca.ulaval.glo4002.game.domain.turn.TurnRepository;
 
+import javax.inject.Inject;
 import java.util.List;
-import java.util.UUID;
-import static ca.ulaval.glo4002.game.domain.resources.ResourceTypesEnum.*;
 
 public class TurnUseCase {
 
     private final TurnFactory turnFactory;
     private final TurnRepository turnRepository;
     private final ResourceRepository resourceRepository;
-    private final TurnAssembler turnAssembler;
     private final ActionRepository actionRepository;
     private final DinosaurRepository dinosaurRepository;
+    private final ZooManager zooManager;
 
     public TurnUseCase(
             TurnFactory turnFactory,
             TurnRepository turnRepository,
             ResourceRepository resourceRepository,
             DinosaurRepository dinosaurRepository,
-            TurnAssembler turnAssembler,
-            ActionRepository actionRepository) {
+            ActionRepository actionRepository,
+            ZooManager zooManager) {
         this.turnFactory = turnFactory;
         this.turnRepository = turnRepository;
         this.resourceRepository = resourceRepository;
         this.dinosaurRepository = dinosaurRepository;
-        this.turnAssembler = turnAssembler;
         this.actionRepository = actionRepository;
+        this.zooManager = zooManager;
     }
 
     public void createTurn() {
@@ -55,67 +52,42 @@ public class TurnUseCase {
         turnRepository.save(turn);
     }
 
-    private void cookIt() {
+    protected void cookIt() {
         resourceRepository.add(new Burger(100));
         resourceRepository.add(new Salad(250));
         resourceRepository.add(new Water(10000));
     }
 
-    public void postAction() {
+    private void postAction() {
         resourceRepository.decreaseExpirationDate();
         feedDinosaurs();
         removeBabyDinosaurs();
     }
 
-    public TurnDto getFromId(UUID id) {
-        Turn turn = turnRepository.findById(id);
-        return turnAssembler.toDto(turn);
-    }
-
-    public void feedDinosaurs() {
+    protected void feedDinosaurs() {
         feedDinosaursByDietType(dinosaurRepository.getSortedDinosaursByStrengthThenName());
     }
 
-    public void removeBabyDinosaurs() {
+    protected void removeBabyDinosaurs() {
         removeBabyDinosaur(dinosaurRepository.getSortedDinosaursByStrengthThenName());
     }
 
     //TODO: Test uniter cette méthode
-    private void feedDinosaursByDietType(List<Dinosaur> sortedDinosaursByStrengthThenName) {
-        boolean dead = false;
+    protected void feedDinosaursByDietType(List<Dinosaur> sortedDinosaursByStrengthThenName) {
+        boolean isStarving;
 
         for (Dinosaur dinosaur : sortedDinosaursByStrengthThenName) {
-                if (dinosaur.getDiet().equals(DietType.HERBIVORE)) {
-                    if (!resourceRepository.removeResources(Salad, dinosaur.feedFood())) {
-                        dead = true;
-                    }
-                } else if (!resourceRepository.removeResources(Burger, dinosaur.feedFood())){
-                        dead = true;
-                }
-
-                if (!resourceRepository.removeResources(Water, dinosaur.feedWater())) {
-                    dead = true;
-                }
-
-                if (dead) dinosaurRepository.remove(dinosaur);
+            isStarving = zooManager.feedThenCheckIfStarving(resourceRepository.getPantry(),dinosaur);
+            if (isStarving) dinosaurRepository.remove(dinosaur);
         }
     }
+
     private void removeBabyDinosaur(List<Dinosaur> sortedDinosaursByStrengthThenName) {
-        boolean dead = false;
         for (Dinosaur dinosaur : sortedDinosaursByStrengthThenName) {
-            if (dinosaur instanceof DinosaurBaby) {
-                if (dinosaurRepository.findByName(((DinosaurBaby) dinosaur).getFatherName()) == null
-                        && dinosaurRepository.findByName(((DinosaurBaby) dinosaur).getMotherName()) == null) {
-                    dead = true;
-                }
-
-                if (dead) {
-                    dinosaurRepository.remove(dinosaur);
-                }
+            if (dinosaur instanceof DinosaurBaby && dinosaurRepository.areBothParentsDead(dinosaur)) {
+                dinosaurRepository.remove(dinosaur);
             }
-    }
-    //Delete baby dinosaur if both parents are dead
-
+        }
     }
 
     public void reset() {
